2024.15

+++ src/_2022/Day09/Part01.cs
@@ -10,7 +10,7 @@ public partial class Part01 : PuzzleSolver<long>
             .SplitLines()
             .Select(_ => _.Split(" ") switch { var x => (dir: x[0], dist: int.Parse(x[1])) });
 
-        var visited = new HashSet<(int, int)>();
+        var visited = new HashSet<(long, long)>();
         var knotPos = new Point[2];
         visited.Add(knotPos[^1]);
