+++ src/_2022/Day24/Part01.cs
@@ -102,7 +102,7 @@ public partial class Part01 : PuzzleSolver<long>
                     .Select(pos => pos + state.Pos)
                     .Where(pos => !newBlizzardPos.Select(b => b.Pos).Contains(pos))
                     .Where(pos => pos == map.EndPos || pos.X >= 1 && pos.X <= map.Width - 1 && pos.Y >= 1 && pos.Y <= map.Height - 1)
-                    .Where(pos => map[pos.Y][pos.X] != '#')
+                    .Where(pos => map[(int)pos.Y][(int)pos.X] != '#')
                     .ToList();
 
                 if (!newBlizzardPos.Any(_ => _.Pos == state.Pos))
